Renamed two files
The filenames now reflect that the interface elements have U and Pw degrees of freedom.

# Conflicts:
#	applications/GeoMechanicsApplication/custom_utilities/extrapolation_utilities.cpp
