+++ applications/GeoMechanicsApplication/custom_utilities/element_utilities.cpp
@@ -11,7 +11,7 @@
 //
 
 #include "element_utilities.hpp"
-#include "custom_elements/interface_element.h"
+#include "custom_elements/U_Pw_interface_element.h"
 
 #include <cstddef>
