+++ applications/GeoMechanicsApplication/custom_utilities/extrapolation_utilities.cpp
@@ -13,7 +13,7 @@
 
 #include "custom_utilities/extrapolation_utilities.h"
 
-#include "custom_elements/interface_element.h"
+#include "custom_elements/U_Pw_interface_element.h"
 #include "custom_utilities/generic_utilities.hpp"
 #include "custom_utilities/linear_nodal_extrapolator.h"
 #include "element_utilities.hpp"
