+++ applications/GeoMechanicsApplication/geo_mechanics_application.h
@@ -83,6 +83,7 @@
 
 // elements
 #include "custom_elements/Pw_element.h"
+#include "custom_elements/U_Pw_interface_element.h"
 #include "custom_elements/U_Pw_small_strain_FIC_element.hpp"
 #include "custom_elements/U_Pw_small_strain_element.hpp"
 #include "custom_elements/U_Pw_small_strain_interface_element.hpp"
@@ -90,7 +91,6 @@
 #include "custom_elements/U_Pw_updated_lagrangian_element.hpp"
 #include "custom_elements/calculation_contribution.h"
 #include "custom_elements/geo_steady_state_Pw_piping_element.h"
-#include "custom_elements/interface_element.h"
 #include "custom_elements/small_strain_U_Pw_diff_order_element.hpp"
 #include "custom_elements/steady_state_Pw_element.hpp"
 #include "custom_elements/transient_Pw_element.hpp"
