+++ applications/GeoMechanicsApplication/test_setup_utilities/element_setup_utilities.cpp
@@ -15,9 +15,9 @@
 #include "custom_conditions/U_Pw_normal_face_load_condition.h"
 #include "custom_conditions/line_load_2D_diff_order_condition.h"
 #include "custom_elements/Pw_element.h"
+#include "custom_elements/U_Pw_interface_element.h"
 #include "custom_elements/U_Pw_small_strain_element.hpp"
 #include "custom_elements/calculation_contribution.h"
-#include "custom_elements/interface_element.h"
 #include "custom_elements/interface_stress_state.h"
 #include "custom_elements/plane_strain_stress_state.h"
 #include "custom_elements/small_strain_U_Pw_diff_order_element.hpp"
